+++ spec/sidekiq/tasks/web_spec.rb
@@ -59,6 +59,30 @@ RSpec.describe "Sidekiq::Tasks::Web", type: :request do
   end
 
   describe "POST /tasks/:name/enqueue" do
+    before { stub_env("RAILS_ENV", "development") }
+
+    it "enqueues the task with permitted params and redirects to the task", :aggregate_failures do
+      task = build_task(name: "foo:bar", args: ["bar"])
+      expect(Sidekiq::Tasks.tasks).to receive(:find_by!).twice.and_return(task) # twice because of redirect
+      expect(task).to receive(:enqueue).with({"bar" => "baz"})
+
+      post "/tasks/foo-bar/enqueue", {"name" => "foo", "args" => {"bar" => "baz"}, "env_confirmation" => "development"}
+
+      expect(last_response).to be_redirect
+      follow_redirect!
+      expect(last_request.path).to eq("/tasks/foo-bar")
+    end
+
+    it "returns a 400 error when confirm param is invalid", :aggregate_failures do
+      task = build_task(name: "foo:bar", args: ["bar"])
+      expect(task).not_to receive(:enqueue)
+
+      post "/tasks/foo-bar/enqueue", {"env_confirmation" => "an_invalid_confirm"}
+
+      expect(last_response.status).to eq(400)
+      expect(last_response.body).to include("Invalid confirm")
+    end
+
     it "redirects to /tasks when the task is not found", :aggregate_failures do
       expect(Sidekiq::Tasks.tasks).to(
         receive(:find_by!)
@@ -66,22 +90,10 @@ RSpec.describe "Sidekiq::Tasks::Web", type: :request do
           .and_raise(Sidekiq::Tasks::NotFoundError)
       )
 
-      post "/tasks/foo-bar/enqueue"
+      post "/tasks/foo-bar/enqueue", {"env_confirmation" => "development"}
 
       expect(last_response.status).to eq(404)
       expect(last_response.body).to include("Task not found")
-    end
-
-    it "enqueues the task with permitted params and redirects to the task", :aggregate_failures do
-      task = build_task(name: "foo:bar", args: ["bar"])
-      expect(Sidekiq::Tasks.tasks).to receive(:find_by!).twice.and_return(task) # twice because of redirect
-      expect(task).to receive(:enqueue).with({"bar" => "baz"})
-
-      post "/tasks/foo-bar/enqueue", {"name" => "foo", "args" => {"bar" => "baz"}}
-
-      expect(last_response).to be_redirect
-      follow_redirect!
-      expect(last_request.path).to eq("/tasks/foo-bar")
     end
 
     it "returns a 400 error and does not enqueue the task when the params are invalid", :aggregate_failures do
@@ -98,6 +110,7 @@ RSpec.describe "Sidekiq::Tasks::Web", type: :request do
             "an_invalid_param" => "qux",
             "another_invalid_param" => "qux",
           },
+          "env_confirmation" => "development",
         }
       )