+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### [Unreleased]
+
+- Add an environment confirmation input to the task enqueuing form.
+
 ### [0.1.0] - 2025-02-23
 
-- Initial release
+- Initial release.